pyproject.toml
[tool.poetry]
name = "pubmed-paper-fetcher"
version = "0.1.0"
description = "A tool to fetch research papers from PubMed with pharmaceutical/biotech company affiliations"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "pubmed_paper_fetcher"}]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.31.0"
pydantic = "^2.5.2"
typer = "^0.9.0"
rich = "^13.7.0"
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
get-papers-list = "pubmed_paper_fetcher.cli:app" 